requirements.txt
asgiref==3.8.1
boto3==1.34.134
botocore==1.34.134
dj-database-url==2.2.0
Django==5.0.6
django-admin-thumbnails==0.2.9
django-filter==24.2
django-jazzmin==3.0.0
django-storages==1.14.3
gunicorn==22.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
